fix(cli): self-heal better-sqlite3 ABI mismatch; keep native module off the CLI hot path (0.4.2)
The daemon's native sqlite module is compiled against whichever Node ran
the install; when the runtime Node differs (brew node 26 vs npm node 20,
observed on the owner's machine) the addon throws an ABI error. Worse,
commands.ts imported db.js statically, so EVERY engram command crashed,
and the spawned daemon died silently in its log. With capture now on by
default this had to self-heal.

- daemon/preflight.ts: try load → on ABI mismatch run 'npm rebuild
  better-sqlite3' in the owning package (prebuilds make this a download,
  not a compile) → retry once → clear manual fix if it still fails
- worker.ts is now an import-light bootstrap that preflights before
  importing the worker body (worker-main.ts) with its static imports
- daemon status loads the DB lazily and degrades gracefully

Verified: CLI startup no longer loads better-sqlite3 at all.
