+++ apps/cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@getengram/cli",
-  "version": "0.4.1",
+  "version": "0.4.2",
   "description": "CLI for Engram \u2014 persistent memory for AI agents",
   "type": "module",
   "bin": {
